requirements.txt
apache-airflow==1.9
xmljson
lxml<4.0,>=3.6.0